requirements.txt
Flask==2.3.3
firebase-admin==6.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7
openai==1.3.5
gunicorn==21.2.0
pytz==2023.3
Flask-Login==0.6.2
Pillow==10.0.0
requests==2.31.0
python-dateutil==2.8.2
cryptography==41.0.3
email-validator==2.0.0.post2